create_kadai

+++ app/Http/Controllers/TasksController.php
@@ -11,24 +11,44 @@ class TasksController extends Controller
     // getでtasks/にアクセスされた場合の「一覧表示処理」
     public function index()
     {
-        //task一覧を取得
-        $tasks = Task::all();
         
-        //task一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // 他人のタスクにアクセスしようとした場合、トップページにリダイレクトさせる条件分岐。
+        if(\Auth::check()) {
+            
+            // 認証済みのユーザを取得
+            $user = \Auth::user();
+       
+            // このユーザのみtaskを取得する。
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(1);
+            
+            //task一覧ビューでそれを表示
+            return view('tasks.index', [
+                'tasks' => $tasks,
+            ]);
+            
+        }else{    
+            
+            return redirect('/');
+        }
+
     }
 
     // getでtasks/createにアクセスされた場合の「新規登録画面表示処理」
     public function create()
     {
-        $task = new Task;
-        
-        //タスク作成ビューを表示
-        return view('tasks.create', [
-            'task' => $task,
-        ]);
+        if(\Auth::id() == $task->user_id){
+            
+            $task = new Task;
+            
+            //タスク作成ビューを表示
+            return view('tasks.create', [
+                'task' => $task,
+            ]);
+            
+        }else{
+            
+            return redirect ('/');
+        }
     }
 
     // postでtasks/にアクセスされた場合の「新規登録処理」
@@ -41,10 +61,17 @@ class TasksController extends Controller
         ]);
         
         //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        // $task = new Task;
+        // $task->status = $request->status;
+        // $task->content = $request->content;
+        // $task->user_id = $request->user_id;
+        // $task->save();
+        
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値を元に作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            ]);
         
         //トップページにリダイレクト
         return redirect('/');
@@ -56,53 +83,82 @@ class TasksController extends Controller
         //idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        //タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        if(\Auth::id() == $task->user_id) {
+            
+            //タスク詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }else{
+            
+            return redirect('/');
+        }
     }
 
     // getでtasks/（任意のid）/editにアクセスされた場合の「更新画面表示処理」
     public function edit($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
+            //idの値でタスクを検索して取得
+            $task = Task::findOrFail($id);
         
-        //タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if(\Auth::id() == $task->user_id) {
+            
+            //タスク編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+        }else{
+            
+            return redirect('/');
+        }
     }
 
     // putまたはpatchでtasks/（任意のid）にアクセスされた場合の「更新処理」
     public function update(Request $request, $id)
     {
-        // バリデーション
-        $request->validate([
-            'status' => 'required|max:10',
-            'content' => 'required|max:255',
-        ]);
+        $task = \App\Tasks::findOrFail($id);
         
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        //タスクを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        if(\Auth::id() == $task->user_id) {
+            
+            // バリデーション
+            $request->validate([
+                'status' => 'required|max:10',
+                'content' => 'required|max:255',
+            ]);
+            
+            //idの値でタスクを検索して取得
+            $task = Task::findOrFail($id);
+            //タスクを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
         
-        //トップページにリダイレクト
-        return redirect('/');
+        }else{
+            
+            //トップページにリダイレクト
+            return redirect('/');
+        }
     }
 
     // deleteでtasks/（任意のid）にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        //タスクを削除
-        $task->delete();
+        $task = \App\Task::findOrFail($id);
         
-        // トップページにリダイレクト
-        return redirect('/');
+        if(\Auth::id() == $task->user_id) {
+            
+            //idの値でタスクを検索して取得
+            $task = Task::findOrFail($id);
+            //タスクを削除
+            $task->delete();
+            
+            return redirect('/');
+            
+        }else{
+            // トップページにリダイレクト
+            return redirect('/');
+            
+        }
+        
     }
 }
